+++ CHANGELOG.md
@@ -20,5 +20,11 @@ Here is a template for new release sections
 
 ### Added
 - CHANGELOG.md
+- geojson serializer of the data from the oep server
 
 ### Changed
+-
+
+### Removed
+- db_sqla.py file
+- wam_environment.yml file